requirements.txt
asgiref==3.5.2
cloudinary==1.29.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.13
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap4==22.1
django-crispy-forms==1.14.0
django-extra-views==0.14.0
gunicorn==20.1.0
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
